Clean up gcc 2.8.0 -Wall warnings... so we don't have to deal with PRs about them.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/1.3.x@80196 13f79535-47bb-0310-9956-ffa450edef68

+++ src/modules/standard/mod_negotiation.c
@@ -1895,7 +1895,7 @@ static void set_neg_headers(request_rec *r, negotiation_state *neg,
                 vary_by_charset = 1;
             }
         }
-        if ((len = (long) find_content_length(neg, variant)) != 0) {
+        if ((len = find_content_length(neg, variant)) != 0) {
             ap_snprintf(lenstr, sizeof(lenstr), "%ld", len);
             rec = pstrcat(r->pool, rec, " {length ", lenstr, "}", NULL);
         }
